refactor: réorganise la structure des fichiers selon les bonnes pratiques (#2)
- Déplace les modules cœur dans src/ (gedcstruct.js, g7datatypes.js, g7lookups.js, g7structure.js)
- Renomme demo/ en examples/ pour plus de clarté
- Met à jour les chemins d'import dans simpleValidator.js et examples/test_findOrCreate.js
- Ajoute package.json avec les exports ES module et les métadonnées du projet
- Ajoute .gitignore
